# openapi.yml
openapi: 3.1.0
info:
  title: DocAI OCR API
  version: 1.0.0
  description: |
    REST API that accepts a PDF and returns a searchable PDF using GCP Document AI
    via a custom ocrmypdf plugin. Includes an endpoint to store results in Google Drive.
  contact:
    name: DocAI OCR API
servers:
  - url: http://localhost:8000
    description: Local development
paths:
  /healthz:
    get:
      operationId: healthCheck
      summary: Health check
      tags:
        - System
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /v1/ocr:
    post:
      operationId: ocrPdf
      summary: OCR a PDF and return searchable PDF
      description: |
        Accepts a PDF file via multipart upload, runs OCR using GCP Document AI
        through ocrmypdf, and returns the same PDF with an embedded searchable text layer.
      tags:
        - OCR
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: PDF file to OCR
      responses:
        "200":
          description: Searchable PDF (PDF/A)
          content:
            application/pdf:
              schema:
                type: string
                format: binary
          headers:
            Content-Disposition:
              schema:
                type: string
              description: Attachment filename for the searchable PDF
        "400":
          $ref: "#/components/responses/BadRequest"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalError"
        "502":
          $ref: "#/components/responses/BadGateway"
  /v1/ocr/drive:
    post:
      operationId: ocrPdfToDrive
      summary: OCR a PDF and upload to Google Drive
      description: |
        Accepts a PDF file, runs OCR, uploads the searchable PDF to a configured
        Google Drive Shared Drive folder, and returns file metadata.
      tags:
        - OCR
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: PDF file to OCR
                filename:
                  type: string
                  description: Optional output filename in Drive (defaults to original name with _ocr suffix)
                  example: searchable-document.pdf
                folder_id:
                  type: string
                  description: Optional Shared Drive folder ID (overrides server default)
                  example: 1abcDEFghijklmnop
      responses:
        "200":
          description: Drive upload metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriveUploadResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalError"
        "502":
          $ref: "#/components/responses/BadGateway"
  /v1/jobs:
    post:
      operationId: createOcrJob
      summary: Submit a PDF for async OCR
      description: |
        Accepts a PDF via multipart upload, streams it to disk, and enqueues an OCR job.
        Returns immediately with a job ID. Poll `GET /v1/jobs/{jobId}` for status and
        download the result from `GET /v1/jobs/{jobId}/result` when status is `succeeded`.
        Supports large documents (1000+ pages, up to configured upload size limit).
      tags:
        - Jobs
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: PDF file to OCR
                filename:
                  type: string
                  description: Optional output filename when downloading the result
                  example: searchable-document.pdf
                folder_id:
                  type: string
                  description: Optional Google Drive folder ID to upload the result on completion
                  example: 1abcDEFghijklmnop
      responses:
        "202":
          description: Job accepted and queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobCreatedResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/jobs/{jobId}:
    get:
      operationId: getOcrJobStatus
      summary: Get OCR job status
      tags:
        - Jobs
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Job identifier returned by POST /v1/jobs
      responses:
        "200":
          description: Current job status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobStatusResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/jobs/{jobId}/result:
    get:
      operationId: getOcrJobResult
      summary: Download OCR job result
      tags:
        - Jobs
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Job identifier returned by POST /v1/jobs
      responses:
        "200":
          description: Searchable PDF output
          content:
            application/pdf:
              schema:
                type: string
                format: binary
          headers:
            Content-Disposition:
              schema:
                type: string
              description: Attachment filename for the searchable PDF
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  schemas:
    HealthResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - ok
          example: ok
    DriveUploadResponse:
      type: object
      required:
        - fileId
        - name
        - webViewLink
      properties:
        fileId:
          type: string
          description: Google Drive file ID
          example: 1abcDEFghijklmnop
        name:
          type: string
          description: Uploaded file name
          example: document_ocr.pdf
        webViewLink:
          type: string
          format: uri
          description: Link to view the file in Google Drive
          example: https://drive.google.com/file/d/1abc/view
    JobCreatedResponse:
      type: object
      required:
        - jobId
        - status
        - statusUrl
      properties:
        jobId:
          type: string
          format: uuid
          description: Unique job identifier
        status:
          type: string
          enum:
            - queued
          example: queued
        statusUrl:
          type: string
          format: uri
          description: URL to poll job status
          example: http://localhost:8000/v1/jobs/550e8400-e29b-41d4-a716-446655440000
    JobStatusResponse:
      type: object
      required:
        - jobId
        - status
        - createdAt
      properties:
        jobId:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - queued
            - running
            - succeeded
            - failed
        pageCount:
          type: integer
          nullable: true
          description: Number of pages processed when succeeded
        error:
          type: string
          nullable: true
          description: Error detail when status is failed
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
          nullable: true
        finishedAt:
          type: string
          format: date-time
          nullable: true
        resultUrl:
          type: string
          format: uri
          nullable: true
          description: Present when status is succeeded
        driveFileId:
          type: string
          nullable: true
        driveWebViewLink:
          type: string
          format: uri
          nullable: true
    ProblemDetails:
      type: object
      description: RFC 7807 Problem Details
      required:
        - type
        - title
        - status
        - detail
      properties:
        type:
          type: string
          format: uri
          example: about:blank
        title:
          type: string
          example: Bad Request
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: Uploaded file must be a PDF
        instance:
          type: string
          format: uri
          example: /v1/ocr
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    PayloadTooLarge:
      description: File exceeds size limit
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    UnprocessableEntity:
      description: PDF could not be processed
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    InternalError:
      description: Internal server error
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    BadGateway:
      description: Upstream service error (Document AI or Drive)
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"